Add ability for server to listen to blockchain
