+++ package.json
@@ -23,6 +23,7 @@
     "dotenv": "^10.0.0",
     "ethers": "^5.5.2",
     "express": "^4.17.1",
+    "hardhat": "^2.7.1",
     "helmet": "^4.6.0",
     "morgan": "^1.10.0",
     "nodemailer": "^6.6.5"
